+++ format.c
@@ -400,6 +400,7 @@ static void fmt_msg_output_str (FILE *fp,
                                         char *json = kc_avro_to_json(
                                                 rkmessage->key,
                                                 rkmessage->key_len,
+                                                NULL,
                                                 errstr, sizeof(errstr));
 
                                         if (!json) {
@@ -446,8 +447,8 @@ static void fmt_msg_output_str (FILE *fp,
                                         char *json = kc_avro_to_json(
                                                 rkmessage->payload,
                                                 rkmessage->len,
-                                                errstr,
-                                                sizeof(errstr));
+                                                NULL,
+                                                errstr, sizeof(errstr));
 
                                         if (!json) {
                                                 what_failed =